requirements.txt
aiohttp
pandas
python-benedict
openpyxl